Backport r19013 from trunk
Thanks to Sally Fortin for drawing attention to this bug, wherein new
Holding Subfields can't be created in this interface via the staff client.

Acq: Tracked down a small but subtle bug with the Holding Subfield tab of
the Provider editor interface.

We apparently have a global variable "name" defined somewhere which, when
assigned in place of an overrideWidget for an AutoGrid thanks to some vestigial
code, ultimately causes an error from within AutoFieldWidget, if you're using
the staff client.  That's because name is the string "browser_browser" in the
staff client, but if you access this interface from Firefox, name is apparently
set to some value that evaluates to false in a boolean context, because
AutoFieldWidget doesn't choke on it (the comparison at line 182 of that file
will return false).



git-svn-id: svn://svn.open-ils.org/ILS/branches/rel_2_0@19014 dcc99617-32d9-48b4-a31d-7c20da2025e4