Add an "Apply full authority" option to the MARC editor
By definition, the only subfields that should be applied from an
authority record to a bibliographic record are the complete set
from the 1XX field of the authority record. So, rather than forcing 
the user to select the correct subfields from the chosen authority
record, give them this handy shortcut to do the right thing.

Also, add copyright / licensing header - hopefully accurate.


git-svn-id: svn://svn.open-ils.org/ILS/branches/rel_2_0@19024 dcc99617-32d9-48b4-a31d-7c20da2025e4