Propagate merged authority record change to bibliographic record
Launchpad bug 688015 reported that the merge of authority records
resulted in the control number identifier in a controlled field
in a bibliographic record changing to the right value, but the
value of the field itself wasn't being changed.

The problem turned out to be a confusion of the "source" and
"target" values in the authority.merge_record() function. Once
that was straightened out, the merge updated the controlled
field as expected.


git-svn-id: svn://svn.open-ils.org/ILS/branches/rel_2_0@19033 dcc99617-32d9-48b4-a31d-7c20da2025e4