Create authority records with meaningful cataloging agency values
Hardcoded values in the 040 are problematic for sites that want
to be a source of authority records; with the addition of another
parameter to the open-ils.cat.authority.record.create_from_bib(.readonly)
API, we can solve that problem. The callers of this API (the MARC editor)
have access to the cat.marc_control_number_identifier OU setting, so as
long as that is set correctly, problem solved.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@19073 dcc99617-32d9-48b4-a31d-7c20da2025e4