Towards a more useful authority delete confirmation pane
Rather than simply creating a raw confirm dialogue that only
provides the cataloguer with the ID of the authority record that
is about to be deleted (not an especially useful handle), supply
a confirmation dialogue that recaps how many bibliographic records
are linked to this authority record and also displays the MARC of
the record about to be deleted.

Most authority records appear to be short enough that the MARC
display is not a concern, but if it becomes a problem we can opt
not to show the MARC by default and give the cataloguer the ability
to click a "View MARC" button on the confirmation dialog instead.

Along the way, make some more variables local and remove some
unnecessary workarounds for object IDs.


git-svn-id: svn://svn.open-ils.org/ILS/branches/rel_2_0@19074 dcc99617-32d9-48b4-a31d-7c20da2025e4