Enhance usability of authority record deletion dialogue
Don't show the number of linked bib records unless it is
a non-zero sum.

Make "Cancel" the button with focus in the dialog, so that
an errant press of the "Enter" key doesn't do the wrong thing.

Trim space from the authority text in the display. Move the
authority text from the dialog title to the contents of the
dialog.

Do not display the MARC for the authority record by default;
add a "View MARC" button and toggle it to show/hide MARC.


git-svn-id: svn://svn.open-ils.org/ILS/branches/rel_2_0@19075 dcc99617-32d9-48b4-a31d-7c20da2025e4