Invoke MARC editor with the appropriate control number identifier
The control number identifier is most critical for the creation
of authority records; given that the callers have the required
context, have them pass in the control number identifier to the
MARC editor. Edit MARC record, new MARC record, import MARC record;
those have to be the big ones, right?


git-svn-id: svn://svn.open-ils.org/ILS/branches/rel_2_0@19076 dcc99617-32d9-48b4-a31d-7c20da2025e4