Clean up marc_export and teach it how to export authority records
The new "--type" parameter enables users to specify "authority"
and have the expected result.

Some other refactoring and fixes:

Two large chunks of code were factored out into subroutines to
make it a little easier to read the main flow of the code.

As we explicitly call uc() on the $format input parameter
to begin with, we don't have to call it subsequently.

If a requested ID did not exist in the database, then
the script would die; now we trap and flag that error but
continue to fulfill subsequent requests.

The --money parameter is now documented in the help, and
the help has a brief overview that describes (poorly) the
expected format for the list of IDs.

An uninitialized variable error when exporting BRE format
records was squashed.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@19083 dcc99617-32d9-48b4-a31d-7c20da2025e4