Create and use authority browse feeds that match against see and see also references
Approaching something workable for recognizing 4xx/5xx entries in
authority records by registering feeds with ".refs." axes that
search the 4xx/5xx as well as the 1xx, then using the feeds in
the MARC editor authority context menu.

Also tighten up the layout of the see / see from entries when
they appear in the authority context menu; "text-align: right" was
a bit over the top.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@19092 dcc99617-32d9-48b4-a31d-7c20da2025e4