Serials: provide a way for users to create and modify serial items
Generally, these are created by caption and pattern-based prediction, and
modified by receiving.  However, some circumstances will warrant manually
creating items, and there need to be significant bumper rails around that
process, as it's otherwise easy to make items that don't respect certain
assumptions of the serials logic (you can still accomplish this if you're
trying, but it's harder to do than it would be if you had the anarchy of
a more straightforward AutoGrid interface).

Access the new interface by clicking any link in the Label column of the
Issuances tab of the Alternate Serials Control Subscription Details inteface.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@19125 dcc99617-32d9-48b4-a31d-7c20da2025e4