Teach supercat browse / startwith to return a distinct set of records
The addition of the .refs variants exposed feeds that had multiple
copies of the same bib in the same results (e.g. if 1xx was "Scott, Daniel"
and 4xx in the same record was "Scott, Dan" it would be returned twice).
Filtering it in this call prevents the callers from requesting 10 records
and only getting 7 back (or whatever). If there's an actual use case for
getting the same record twice, we can revisit this.

Note that the authority context menu currently still shows duplicate
records, because the record is displayed in full for each 4xx / 5xx
entry, that will be fixed in a subsequent patch.

Also, prevent supercat browse from issuing two separate requests for
page = 0; the second would always stomp on the results of the first,
so I just made the first request go away in that case.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@19131 dcc99617-32d9-48b4-a31d-7c20da2025e4