Backport r19137 from trunk
Make openils.PermaCrud.Store use a default order_by clause in common situations

This avoids an issue where, using a FilteringSelect dijit, sometimes you type a
couple letters, let's says "ABC", and you get a set of matches like this:

"ABCDE"
"ABCD"
"ABC"

just because there happens to be no specific ordering.  If you click the ABC
match and tab away too quickly, the dijit will reset itself to the first match,
"ABCDE".  This is probably a Dojo/Dijit bug, as I've been able to duplicate the
problem with other stores, but that's all murky.

This avoids the issue by making sure the shortest match is at the top of the
list.



git-svn-id: svn://svn.open-ils.org/ILS/branches/rel_2_0@19138 dcc99617-32d9-48b4-a31d-7c20da2025e4