Simplify authority reference deduplication
Rather than multiplying the number of records to be returned in
the case of a .refs. call, stick with the original number. This
avoids corner cases where paging through the results would skip
some records, and the UI is not overly heavily affected as even
in the case where one record offers 10 matches for see also /
see from references and represents an entire page of results,
the UI will display 11 lines (1 for the 1XX field and 10 for
the 4XX / 5XX fields). The user will be able to page to the next
set of results to continue browsing.

We can also be smarter about some things:
  * Return the original list immediately if not a .refs. call
  * No need to reverse and unshift the list items when
    deduping the .refs. list; just stick with the same
    order of the incoming list


git-svn-id: svn://svn.open-ils.org/ILS/trunk@19144 dcc99617-32d9-48b4-a31d-7c20da2025e4