Serials: a regularity (i.e. 85X subfield $y) page for the caption/pattern wizard
This adds a new page to the caption/pattern wizard that allows the user to
enter regularity information (specific published, omitted, and combined issues)
by chronology.  Doing the same by enumeration is possible in MFHD but not yet
supported in the wizard.

You still have to be a serials librarian who understands MFHD and the 85X tags
in order to really benefit from this, but it beats hand-entering the MARC tags.

Still to-do to perfect this:
    -   suggest (pre-enable) the regularity page when numeric $w is used
    -   use grid layout instead of hbox and vbox elements for neatness
    -   support enumeration codes
    -   days of month widget should be smarter than to always allow 31 days
    -   the whole caption/pattern wizard still needs scrollbars
    -   more limitations to prevent the user from entering patterns that
            don't make sense
    -   make sure that if a user fills out a page of the wizard, including
            this new one, but then unchecks the whole page, that whatever
            work they did is not included in the compiled result
    -   more testing, general cleanup


git-svn-id: svn://svn.open-ils.org/ILS/trunk@19181 dcc99617-32d9-48b4-a31d-7c20da2025e4