Make EditPane objects built of AutoFieldWidget objects (such as those used for create/edit dialogs with AutoGrid) enforce required fields more forcefully.
Before, if a field was marked required either in the IDL or by the
requirdFields attribute of an AutoGrid, you'd get a yellow widget with a caution
sign for that field, but you could still click save and the system would
attempt to save your object.

Sometimes this is stopped when pcrud can't save the object due to
required="true" in the IDL and/or a "not null" constraint in the schema, but
there may be cases where a given interface wants to require a value in a given
field even though that's not necessarily enforced at lower levels.

Serials: Specifically use this new feature in the distribution pane of the
Alternate Serial Control view, to prevent the creation of issues without a
"receive unit template" field, as you can't receive items in the Batch
Receive interface without one.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@19192 dcc99617-32d9-48b4-a31d-7c20da2025e4