Build and flatten a tree, correcting a sorting issue in some OU dropdowns.
The previous code assumed that work org units would be delivered in hierarchical order, but alas, they are not. Thus, we build the hierarchy and then flatten it, sorting at each level. This will be non-fast with many work OUs, but the common case is a small set, which is not painful.

Further improvement is warranted when the above proves false.



git-svn-id: svn://svn.open-ils.org/ILS/branches/rel_2_0@19209 dcc99617-32d9-48b4-a31d-7c20da2025e4