more overzealous unsaved data prompts
onKeypress catches all sorts of stuff, like the control+w for closing tabs.  Not good.  Ideally we'd test event.isChar, but that's broken in Mozilla:
https://bugzilla.mozilla.org/show_bug.cgi?id=312552

Instead we look to see if control, alt, or meta are being held down.  Not perfect, but good enough for now.  For example, won't ignore tabs for jumping from field to field, and won't realize the hotkey for clipboard pasting should trigger the unsaved data state.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@19257 dcc99617-32d9-48b4-a31d-7c20da2025e4