Teach the i18n Makefile how to handle serial.properties
Also check in the POT for serial.properties and some other
foobar.label matches for foobar.accesskey (does not have
an effect on properties files but hopefully if we're
consistent then the pattern will be evident for entities).

Noting in passing that there are a number of "foobar(s)"
comined singular/plural labels that should be split to
support translation.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@19282 dcc99617-32d9-48b4-a31d-7c20da2025e4