Update README to include --admin-user / --admin-pass options for eg_db_config.pl
As the default Evergreen administrator account is now created with
a randomized user name and password, we have added the options to
eg_db_config.pl to reset those to something known and specific to
a given instance.

Also, Fedora 13 is getting close to retirement, update to Fedora 14 for
supported distros.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@19283 dcc99617-32d9-48b4-a31d-7c20da2025e4