For more predictable authority browse results, use startwith until browse is sorted out
authority_tag_sf_browse() was returning unexpected results when only a small
set of authority records was loaded. This may be an indication of a corner
case, but until we sort that out the more straightforward startwith should
be used.

Also, chop the final character of the incoming search value to ensure
that an exact match is returned on page 0 instead of page -1. For example,
'Bacon, Jono.' becomes 'Bacon, Jono' for the search. (The right thing to
do, come to think of it, is probably to naco_normalize the incoming value;
perhaps next patch).


git-svn-id: svn://svn.open-ils.org/ILS/trunk@19331 dcc99617-32d9-48b4-a31d-7c20da2025e4