In actor.usr_purge_data, default destination user to admin (id=1).
Destination user is where we move things like reports and acq picklists when
purging staff members.  However, if we tried to purge a patron that used to be
a staff member, then destination user was being set to null, which won't work
when updating things like requestor on holds.  The purge UI checks for the
STAFF_LOGIN permission on the user to be purged for determining whether to
prompt for a destination user or not.

Some alternatives to this change might include:
1) always asking for a destination user (and defaulting to admin or the
logged in staff member), or
2) robustifying the stored procedure and surrounding code to raise an
expected exception whenever this edge case hits, prompting the UI to try
again after asking for a destination user.


git-svn-id: svn://svn.open-ils.org/ILS/trunk@19344 dcc99617-32d9-48b4-a31d-7c20da2025e4