Two binding fixes for manage items
The first fix is more or less cosmetic. When selecting a target unit, both areas need to refresh in order to not show the selected unit's items in the main list.

The second fix involves handling cases where moving items from one unit to another does not empty the 'donor' unit. In those cases, we need to update the contents of both the new and the old unit.


git-svn-id: svn://svn.open-ils.org/ILS/branches/rel_2_0@19374 dcc99617-32d9-48b4-a31d-7c20da2025e4