JS stubbed out, mostly. See longer commit msg.
Here's the deal with JS so far.

A few instances of onfocus, onblur, onmouseout, and onmouseover have been
left alone, as they're doing simple cosmetic things.

A lot of instances of onclick remain, and a couple onchange, but these
now either a) just do cosmetic things (hide/unhide) or b) call stubbed
out functions prepared in misc.js

That doesn't mean we won't still totally rip out lots of this JS, but
it leaves us a hint as to what is supposed to happen where.  I can
easily see us losing track of that later.  So anyway, many of the
stubbed functions will probably be replaced with purely server side
functionality later, but until then, they can serve as a kind of
documentation.