Ensure new authority ID subfield is inserted in the correct XUL DOM location
Addresses LP 712499. After creating an authority via the context menu
in the MARC editor, the new ID subfield ($0) would be created right after
the subfield on which the context menu was invoked. It turns out that it
was being placed in the wrong location, and one symptom was that the
Validate button would not validate the controlled field against the newly
created authority.

Now we hunt through the parent DOM nodes until we find the 'sf_box' element
and then we append the ID subfield to that node. We also eliminate some
duplicate code by defining a common function so that the problem can be
fixed in one stroke...


git-svn-id: svn://svn.open-ils.org/ILS/trunk@19402 dcc99617-32d9-48b4-a31d-7c20da2025e4