Patch from Thomas Berezansky addressing his recent dynamic circ/hold field weighting patch.
Due to an oversight of auto-casting on my part dynamic weighting may incorrectly order similar rows that differ only in org unit or permission group.

Bad me for not testing for that case before now.

The attached patch should fix this. All it does is replace two functions, and in them replaces one INT variable with a NUMERIC(6,2) and a pile of integer constants with numeric constants (by virtue of adding .0 to them).



git-svn-id: svn://svn.open-ils.org/ILS/branches/rel_2_1@19423 dcc99617-32d9-48b4-a31d-7c20da2025e4