Backport r19470 from trunk
Booking: fix the targeter in the I'm-reserving-this-exact-resource-with-this-
exact-barcode case.

It should have always worked like this. This should better mirror the behavior
in the holds targeter when placing a copy-level hold.  Basically don't fail
to target a resource for a reservation later just because it's not in an
available status right at this moment.



git-svn-id: svn://svn.open-ils.org/ILS/branches/rel_1_6@19473 dcc99617-32d9-48b4-a31d-7c20da2025e4