big ol' commit making advanced search work, albeit with issues
most obvious issue is that advanced search leads to a result page that
renders with a basic search form filled out in a silly-looking way

I'm thinking it'd be best when showing results from an advanced search
to *not* show that basic search form, but instead have a hidden version
of the advanced search form, prepopulated, revealable by a button
labeled "refine your search" or something.

That way we'd not have to worry about reversing query parser syntax back
into broken-out widget values. make sense?