Correct the install location for Dojo i18n bundles
We had been converting xx-YY into xx/yy for a directory structure,
but current evidence shows that it should be xx-yy. Hard to believe
we've overlooked this for so long. Also, with the exception of the
en-ca/en-gb/en-us locales, we should probably trim the locale to
just xx - but that's a different problem, and let's just focus on
getting this working to begin with.

Addresses Launchpad # 723489


git-svn-id: svn://svn.open-ils.org/ILS/branches/rel_2_1@19508 dcc99617-32d9-48b4-a31d-7c20da2025e4