Enable MFHD tests to find their testdata
When the MFHD tests wrapped in mfhd.t are invoked from outside of the directory,
they tests failed to find the mfhddata.txt file. Judicious use of the __FILE__
special literal and File::Basename::dirname() avoids that problem. Now we just
need to munge @INC to ensure that testlib.pm can be resolved (and hope that
no conflicts ever turn up).


git-svn-id: svn://svn.open-ils.org/ILS/branches/rel_2_1@19568 dcc99617-32d9-48b4-a31d-7c20da2025e4