SRU: Restructure the hash returned from open-ils.supercat.biblio.search_aliases
By returning explicit index and title elements from the method,
we can backport cleanly to 2.0 and set the stage for pulling the
title element directly from the database in 2.1 and beyond.

In addition, the enriched object enables us to get rid of the
hardcoded nested qualifier map for bibs, and also enables us to
fix the explain document for SRU for authority records.

Finally, we also fix the mappings of SRU qualifiers to Evergreen
search aliases (in many cases we had been quietly falling back
to plain kw searches, which was not cool).


git-svn-id: svn://svn.open-ils.org/ILS/trunk@19573 dcc99617-32d9-48b4-a31d-7c20da2025e4