Add complete set of Evergreen search aliases and attributes to SRU explain
Original "pull from database" code had limited the search aliases to those
that were qualified; now we pull all of them, and if they are not qualified,
we qualify them as "eg".

In addition, for some of the non-index attributes (sort, direction, site,
available), restore the descriptive title that was previously available.
Slightly longer term we will add a "description" column to
config.metabib_search_alias to store this information.

Also, we had always been using the global %qualifier_map in
return_sru_explain(), even though the intention was to enable the authority
explain index to pass in an overriding set of values. This commit breaks the
authority SRU explain, but as it had been incorrectly dumping the bib explain,
this brings us a step closer to proper explain support for authorities.


git-svn-id: svn://svn.open-ils.org/ILS/branches/rel_2_1@19574 dcc99617-32d9-48b4-a31d-7c20da2025e4