Be more defensive when parsing SFX target results
SFX in the wild has been observed returning targets for getFullTxt
requests that have no target URLs, which makes it rather impossible
to provide a link to the full text resource. Skip such targets (and
for good measure, skip them if they don't have a name either).

Also, change from the SID "conifer" to the more generic and accurate
"evergreen".


git-svn-id: svn://svn.open-ils.org/ILS/trunk@19645 dcc99617-32d9-48b4-a31d-7c20da2025e4