Foundational work for temporary/anon lists and per-user lists (bookbags)
There were already some features in EGCatLoader for adding and
deleting items from lists, but there's more to do, as that code only
dealt with numeric IDs, and we need records avaiable. Also, to avoid
search bots creating temporary/anon lists all the time, and just because
the following is good practice, our "add to my list" links need to be
forms that POST, not links that GET.

Etc., etc.; more to come.