Improve search term propagation from page to page
Also, for things like the author link on the record detail page, build a
link with 'qtype=author&query=JK Rowling' instead of 'query=au:JK Rowling'

The latter works, but causes the search form to look stupid.
Query type dropdown will be set to keyword by default, and the term will
be "au:JK Rowling", which will be a head scratcher for 99% of real
users.