Make DROP statement match new basetype for agg_tsvector
When the CREATE AGGREGATE function was changed to have a basetype
of pg_catalog.agg_tsvector, the corresponding DROP statement was
not changed in a similar fashion; ergo, on creating the schema
a subsequent time, the CREATE statement would fail because the
aggregate function already existed and the transaction would be
rolled back.

Also - do we need to differentiate between PostgreSQL 8.4 and 9.0
here?


git-svn-id: svn://svn.open-ils.org/ILS/trunk@19758 dcc99617-32d9-48b4-a31d-7c20da2025e4